Added segue from Play Area View Controller to Drawing Resesult View Controller.
Changed from one trigger line that is hardcoded to Drawing Pallet View, to two Trigger Line constained to each edge of the screen.  Added hide and unhide function to Trigger line to react accoringly.
Changed Drawing Pallet View so that width is not hard coded.  Instead it is a multiplier to it's super's width.